farm: improve AllRepairs strategy
The previous implementation scattered a goroutine per inconsistent
key-member. If a key was completely empty on a node, for example when a
node is replaced and comes back empty, this can mean tens of thousands
of goroutines for a single multi-key Select.

This implementation serializes Score checking, which takes some time,
but conserves memory. If this proves too slow, we can refactor Scorer to
take a slice of KeyMembers.